+++ pkg/common/OpenGLRenderer.cpp
@@ -178,20 +178,20 @@ void OpenGLRenderer::render(const shared_ptr<Scene>& _scene,Body::id_t selection
 	// set light sources
 	glLightModelf(GL_LIGHT_MODEL_TWO_SIDE,1); // important: do lighting calculations on both sides of polygons
 
-	const GLfloat pos[4]	= {lightPos[0],lightPos[1],lightPos[2],1.0};
+	const GLfloat pos[4]	= {(float) lightPos[0], (float) lightPos[1], (float) lightPos[2],1.0};
 	const GLfloat ambientColor[4]={0.2,0.2,0.2,1.0};
 	const GLfloat specularColor[4]={1,1,1,1.f};
-	const GLfloat diffuseLight[4] = { lightColor[0], lightColor[1], lightColor[2], 1.0f };
+	const GLfloat diffuseLight[4] = { (float) lightColor[0], (float) lightColor[1], (float) lightColor[2], 1.0f };
 	glLightfv(GL_LIGHT0, GL_POSITION,pos);
 	glLightfv(GL_LIGHT0, GL_SPECULAR, specularColor);
 	glLightfv(GL_LIGHT0, GL_AMBIENT, ambientColor);
 	glLightfv(GL_LIGHT0, GL_DIFFUSE, diffuseLight);
 	if (light1) glEnable(GL_LIGHT0); else glDisable(GL_LIGHT0);
 
-	const GLfloat pos2[4]	= {light2Pos[0],light2Pos[1],light2Pos[2],1.0};
+	const GLfloat pos2[4]	= {(float) light2Pos[0], (float) light2Pos[1], (float) light2Pos[2],1.0};
 	const GLfloat ambientColor2[4]={0.0,0.0,0.0,1.0};
 	const GLfloat specularColor2[4]={1,1,0.6,1.f};
-	const GLfloat diffuseLight2[4] = { light2Color[0], light2Color[1], light2Color[2], 1.0f };
+	const GLfloat diffuseLight2[4] = { (float) light2Color[0], (float) light2Color[1], (float) light2Color[2], 1.0f };
 	glLightfv(GL_LIGHT1, GL_POSITION,pos2);
 	glLightfv(GL_LIGHT1, GL_SPECULAR, specularColor2);
 	glLightfv(GL_LIGHT1, GL_AMBIENT, ambientColor2);
